Rename module 'digital' to 'app'
